opt: jump to increase code before check interrupt.
We used to include a copy of inc_pc_jchkint.ds to increase the PC and jump to the code that checks for interrupts. Now we have written a code that increase the PC before the code that check the interrupt, and instructions can jump there to perform the same operation as inc_pc_jchkint.ds.