fix: you can't load opcode in advance!
II got into a big mistake, I shouldn't load the opcode of the next instruction before the check interrupt stage was completed! Consider this case:
CPU is executing arbitrary instruction A, and the next instruction is SJMP. You can extract SJMP opcodes into IR, because IR is now free. But, unfortunately, there is an interrupt request coming in, and the next instruction should be loaded from the interrupt vector! So, SJMP, where are you going to?